More robust method of checking for error
For some reason the API does not append the News24 stream information
if the series name begins with an integer. This change checks whether
the first item in the array looks like News24 info, and applies the
workaround if it does.

This will probably break the plugin if you're actually trying to watch
the News24 stream.