Switch to mobile API
Switch add-on to the android/ios iView API.  The general layout has been
for the most part retained although there are some differences in the
program categories.
I've borrowed the threading code from the plus7 add-on to speed up the
listing of episodes as we now have to fetch each episodes individual
page to get the metadata we want.

Tested on Win 10/Kodi Beta 1 and RPi3/LibreELEC 7.0.2

The only issue I've come up against I'm pretty sure is to do with how
Kodi handles and stores cookies. I've had to clear the cookes.dat file a
couple of times to get rid of HTTP 403 errors. If it ends up being a
recurring issue I can implement manual processing of the m3u8 and
cookies. See this thread on the Kodi forum for a similar example -
[http://forum.kodi.tv/showthread.php?tid=232439&pid=2057557]